+++ OrangeDesignSystem/Sources/OrangeDesignSystem/Components/Image/ODSImage.swift
@@ -8,30 +8,29 @@
 
 import SwiftUI
 
-///
-/// As image can be loaded from resources (sync) or
-/// from url (Async), this desplays image accrding to the
+/// As image can be loaded from resources (sync) or from url (Async), this desplays image according to the
 /// source __ODSImage.Source__.
 ///
 /// For Async image, until the image loads, the view displays a standard placeholder that fills the available space.
 /// For more detail see __AsyncImage__.
-///
+/// Async image can also be cached, meaning if the request to load the image fails some cache if existing ill be used to load the previously fetched data.
 public struct ODSImage: View {
 
-    // ======================
-    // MARK: Store Properties
-    // ======================
+    // =======================
+    // MARK: Stored Properties
+    // =======================
+
+    let source: Source
 
     public enum Source {
         case image(Image)
         case asyncImage(URL, Image)
+        case cachedAsyncImage(URL, Image)
 
         public init(url: URL, placeholder: Image = Image("ods_empty", bundle: Bundle.ods)) {
             self = .asyncImage(url, placeholder)
         }
     }
-
-    let source: Source
 
     // ==========
     // MARK: Body
@@ -39,6 +38,11 @@ public struct ODSImage: View {
 
     public var body: some View {
         switch source {
+        case let .image(image):
+            image
+                .resizable()
+                .aspectRatio(contentMode: .fill)
+
         case let .asyncImage(url, placeHolder):
             AsyncImage(url: url) { image in
                 image
@@ -50,10 +54,16 @@ public struct ODSImage: View {
                     .aspectRatio(contentMode: .fill)
             }
 
-        case let .image(image):
-            image
-                .resizable()
-                .aspectRatio(contentMode: .fill)
+        case let .cachedAsyncImage(url, placeHolder):
+            CachedAsyncImage(url: url) { image in
+                image
+                    .resizable()
+                    .aspectRatio(contentMode: .fill)
+            } placeholder: {
+                placeHolder
+                    .resizable()
+                    .aspectRatio(contentMode: .fill)
+            }
         }
     }
 }